Update Course, fix justfile
