+++ API/project/src/main/java/com/grademanager/demo/service/courseService.java
@@ -1,6 +1,7 @@
 package com.grademanager.demo.service;
 
 import java.util.Optional;
+import org.springframework.http.ResponseEntity;
 import com.grademanager.demo.model.Course;
 import com.grademanager.demo.repository.CourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,12 +17,17 @@ public class CourseService {
     @Autowired
     CourseRepository courseRepository;
 
-    public void createCourse(Course course){
-        courseRepository.save(course);
+    public Course createCourse(Course course){
+        return courseRepository.save(course);
     }
 
-    public Optional<Course> getCourse(Integer id){
-        return courseRepository.findById(id);
+    public Course getCourse(Integer id){
+      Optional <Course> optionalCourse = courseRepository.findById(id);
+        if(!optionalCourse.isPresent()){
+            String err = String.format("The course having ID %s was not found", id);
+            System.out.println(err);
+        }
+        return optionalCourse.get();
     }
 
     public List<Course> getAllCourses(){
@@ -34,7 +40,7 @@ public class CourseService {
         courseRepository.deleteById(id);
     }
 
-    public void updateCourse(Course course){
-        courseRepository.save(course);
+    public Course updateCourse(Course course){
+        return courseRepository.save(course);
     }
 }
